feat: Optionsコンポーネントをリファクタリングし、ActionForm、ActionsList、ExampleSectionコンポーネントを追加して機能を整理
